+++ widgets/Redactor.php
@@ -26,6 +26,7 @@ use yii\helpers\ArrayHelper;
  * Class Redactor
  * @package yii\redactor\widgets
  * @property AssetBundle $assetBundle
+ * @property string $sourcePath
  */
 class Redactor extends InputWidget
 {
@@ -52,7 +53,10 @@ class Redactor extends InputWidget
         }
     }
 
-    public function defaultOptions()
+    /**
+     * Sets default options
+     */
+    protected function defaultOptions()
     {
         if (!isset($this->options['id'])) {
             if ($this->hasModel()) {
@@ -61,24 +65,28 @@ class Redactor extends InputWidget
                 $this->options['id'] = $this->getId();
             }
         }
-        $this->clientOptions['imageUpload'] = Url::to(ArrayHelper::getValue($this->clientOptions, 'imageUpload', Yii::$app->getModule('redactor')->imageUploadRoute));
-        $this->clientOptions['fileUpload'] = Url::to(ArrayHelper::getValue($this->clientOptions, 'fileUpload', Yii::$app->getModule('redactor')->fileUploadRoute));
+        $this->setOptionsKey('imageUpload', Yii::$app->getModule('redactor')->imageUploadRoute);
+        $this->setOptionsKey('fileUpload', Yii::$app->getModule('redactor')->fileUploadRoute);
+
         $this->clientOptions['imageUploadErrorCallback'] = ArrayHelper::getValue($this->clientOptions, 'imageUploadErrorCallback', new JsExpression("function(json){alert(json.error);}"));
         $this->clientOptions['fileUploadErrorCallback'] = ArrayHelper::getValue($this->clientOptions, 'fileUploadErrorCallback', new JsExpression("function(json){alert(json.error);}"));
 
         if (isset($this->clientOptions['plugins']) && array_search('imagemanager', $this->clientOptions['plugins']) !== false) {
-            $this->clientOptions['imageManagerJson'] = Url::to(ArrayHelper::getValue($this->clientOptions, 'imageManagerJson', Yii::$app->getModule('redactor')->imageManagerJsonRoute));
+            $this->setOptionsKey('imageManagerJson', Yii::$app->getModule('redactor')->imageManagerJsonRoute);
         }
         if (isset($this->clientOptions['plugins']) && array_search('filemanager', $this->clientOptions['plugins']) !== false) {
-            $this->clientOptions['fileManagerJson'] = Url::to(ArrayHelper::getValue($this->clientOptions, 'fileManagerJson', Yii::$app->getModule('redactor')->fileManagerJsonRoute));
+            $this->setOptionsKey('fileManagerJson', Yii::$app->getModule('redactor')->fileManagerJsonRoute);
         }
     }
 
-    public function registerRegional()
+    /**
+     * Register language for Redactor
+     */
+    protected function registerRegional()
     {
-        $lang = ArrayHelper::getValue($this->clientOptions, 'lang', Yii::$app->language);
-        $langAsset = 'lang/' . $lang . '.js';
-        if (file_exists(Yii::getAlias($this->assetBundle->sourcePath . '/' . $langAsset))) {
+        $this->clientOptions['lang'] = ArrayHelper::getValue($this->clientOptions, 'lang', Yii::$app->language);
+        $langAsset = 'lang/' . $this->clientOptions['lang'] . '.js';
+        if (file_exists($this->sourcePath . DIRECTORY_SEPARATOR . $langAsset)) {
             $this->assetBundle->js[] = $langAsset;
         } else {
             ArrayHelper::remove($this->clientOptions, 'lang');
@@ -86,33 +94,45 @@ class Redactor extends InputWidget
 
     }
 
-    public function registerPlugins()
+    /**
+     * Register plugins for Redactor
+     */
+    protected function registerPlugins()
     {
         if (isset($this->clientOptions['plugins']) && count($this->clientOptions['plugins'])) {
             foreach ($this->clientOptions['plugins'] as $plugin) {
                 $js = 'plugins/' . $plugin . '/' . $plugin . '.js';
-                if (file_exists(Yii::getAlias($this->assetBundle->sourcePath . DIRECTORY_SEPARATOR . $js))) {
+                if (file_exists($this->sourcePath . DIRECTORY_SEPARATOR . $js)) {
                     $this->assetBundle->js[] = $js;
                 }
                 $css = 'plugins/' . $plugin . '/' . $plugin . '.css';
-                if (file_exists(Yii::getAlias($this->assetBundle->sourcePath . '/' . $css))) {
+                if (file_exists($this->sourcePath . DIRECTORY_SEPARATOR . $css)) {
                     $this->assetBundle->css[] = $css;
                 }
             }
         }
     }
 
-    public function registerScript()
+    /**
+     * Register clients script to View
+     */
+    protected function registerScript()
     {
         $clientOptions = (count($this->clientOptions)) ? Json::encode($this->clientOptions) : '';
         $this->getView()->registerJs("jQuery('#{$this->options['id']}').redactor({$clientOptions});");
     }
 
-    public function registerAssetBundle()
+    /**
+     * Register assetBundle
+     */
+    protected function registerAssetBundle()
     {
         $this->_assetBundle = RedactorAsset::register($this->getView());
     }
 
+    /**
+     * @return AssetBundle
+     */
     public function getAssetBundle()
     {
         if (!($this->_assetBundle instanceof AssetBundle)) {
@@ -121,4 +141,20 @@ class Redactor extends InputWidget
         return $this->_assetBundle;
     }
 
+    /**
+     * @return bool|string The path of assetBundle
+     */
+    public function getSourcePath()
+    {
+        return Yii::getAlias($this->getAssetBundle()->sourcePath);
+    }
+
+    /**
+     * @param $key
+     * @param null $defaultValue
+     */
+    protected function setOptionsKey($key, $defaultValue = null)
+    {
+        $this->clientOptions[$key] = Url::to(ArrayHelper::getValue($this->clientOptions, $key, $defaultValue));
+    }
 }